% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{id}
\alias{id}
\title{Sampled Fasta file of chromosome sequence from hg19 blacklist}
\format{A character sequence with 20 elements}
\usage{
data(id)
}
\description{
This dataset is sampled from The hg19 blacklist.
For splitting a chromosome Fasta file, sometimes the
Fasta identifier is too complicated to manipulate. This data
can be used to show how to simplify the Fasta identifier.
}
\examples{
data(id)
}
\references{
Satpathy A T, Granja J M, Yost K E, et al. (2019) Nature biotechnology 37,925–936.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/31375813}{PubMed})
}
\keyword{datasets}
