% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceText.R
\name{replaceText}
\alias{replaceText}
\title{Replace tedious chromosome identifier into simple format}
\usage{
replaceText(type = "text", input = input)
}
\arguments{
\item{type}{This can be either "text" or "list", The previous is a large character containing each line of the Fasta file, the latter is a list in which each element contains a unit of Fasta file}

\item{input}{The large character or list containing ids that need to be simplified}
}
\value{
The large character or list of Chromosome Fasta file with simplified id.
}
\description{
Make the chromosome id starting with ">" into simple format like
">chr:1091194-1093520...",this is helpful for sorting the chromosome according to their number
}
\examples{
data("id")
simpleID<- replaceText(type = "text",input = id)
}
\author{
Shaoqian Ma
}
