% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFasID.R
\name{subFasID}
\alias{subFasID}
\title{Subtract chromosome ids from Fasta file}
\usage{
subFasID(text = text)
}
\arguments{
\item{text}{Large character read by readLines from chromosome Fasta file.}
}
\value{
The id list of the Fasta file.
}
\description{
Subtract chromosome ids from Fasta file
}
\examples{
data("text")
text<- replaceText(type = "text",input = text)
id <- subFasID(text = text)
}
