% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tex}
\alias{tex}
\title{Fasta file of chromosome sequence produced from sequence character}
\format{A large list containing 62 elements.}
\usage{
data(tex)
}
\description{
Data from "Three representative inter and intra-subspecific crosses
 reveal the genetic architecture of reproductive isolation in rice."
}
\examples{
data(tex)
}
\references{
Li, G. et al. (2017) The Plant Journal 92, 349–362.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28805257}{PubMed})
}
\keyword{datasets}
