% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{text}
\alias{text}
\title{Fasta file of chromosome sequence}
\format{A character sequence with 2099 elements.}
\usage{
data(text)
}
\description{
A downsampled dataset containing the hg19 chromosome sequence from
the hg19 blacklist. The hg19 blacklist is obtained from the supplementary
dataset from "Massively parallel single-cell chromatin landscapes of human 
immune cell development and intratumoral T cell exhaustion."
The dataset is sent to the UCSC Table Browser for obtaining the corresponding 
sequence file. The sequence file is processed with replaceText function to 
simplify the fasta id. To best illustate the usage, the sequence file is downsampled.
}
\examples{
data(text)
}
\references{
Satpathy A T, Granja J M, Yost K E, et al. (2019) Nature biotechnology 37, 925–936.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/31375813}{PubMed})
}
\keyword{datasets}
